% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_imputation.R
\name{surv_imputation}
\alias{surv_imputation}
\title{preprocessing test and train for KNN imputation}
\usage{
surv_imputation(exp_dat, time_var, target_var)
}
\arguments{
\item{exp_dat}{is a list of the train and test dataframes, an external dataset can also be included.}

\item{time_var}{is the column in the df corresponding to time.}

\item{target_var}{is the column in the df corresponding to the target variable.}
}
\value{
It returns a list containing train and test.
}
\description{
This function takes train and test dfs and replaces NA values with KNN imputation
with a user specified value for K
}
