% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_model.R
\name{surv_model}
\alias{surv_model}
\title{Survival ML modelling}
\usage{
surv_model(exp_dat, modType, time, target, ids, grid)
}
\arguments{
\item{exp_dat}{is the list of training and test dfs}

\item{modType}{is the type of model required}

\item{time}{is the time column in the df}

\item{target}{is the target column in the df}

\item{ids}{is a vector of two strings used for the model ids.}

\item{grid}{is the hyperparameter tuning grid in a list.}
}
\value{
it saves a model and returns a vector of C-Indexes for train and test.
}
\description{
This function does the actual survival based ML modelling.
}
