% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_splitter.R
\name{dat_splitter}
\alias{dat_splitter}
\title{Reading in dataframe and returning a list containing train and test dfs}
\usage{
dat_splitter(dat, perc, target)
}
\arguments{
\item{dat}{is ...}

\item{perc}{is ...}

\item{target}{is ...}
}
\value{
it returns a list containing two dataframes, one train, one test
}
\description{
This function ...
}
